create progressive list of terms for data
each reactive term takes the last selected category by the user and outputs the list of terms for that category within the respective dataset (i.e. trade, production, landings). This accurately reflects whether the data contains the species selected by incorporating all user selections.

For example, Drums exist in multiple categories (there are freshwater and marine species). Thus, if a dataset contains marine drum data, but not FW drum data, the app would have erroneously stated that there is insufficient data, and not allowed the user to unfilter the data (the button we created). That is now fixed.